+++ Luxelot.Apps.Common/DHT/IBucketEntryValue.cs
@@ -1,4 +1,0 @@
-namespace Luxelot.Apps.Common.DHT;
-
-public interface IBucketEntryValue {
-}