+++ Luxelot.Apps.Common/INode.cs
@@ -1,0 +1,6 @@
+namespace Luxelot.Apps.Common;
+
+public interface INode
+{
+    public event EventHandler<PeerConnectedArgs>? PeerConnected;
+}