+++ apps/Luxelot.Apps.DhtApp/Constants.cs
@@ -1,7 +1,7 @@
-namespace Luxelot.DHT;
+namespace Luxelot.Apps.DHT;
 
 public static class Constants {
     public const int K = 3;
     public const int BUCKET_ENTRY_KEY_BIT_LENGTH = 128;
-    public const int TREE_HEIGHT = Apps.Common.Constants.THUMBPRINT_LEN * 8;
+    public const int TREE_HEIGHT = Common.Constants.THUMBPRINT_LEN * 8;
 }