+++ apps/Luxelot.Apps.DhtApp/IBucketEntryValue.cs
@@ -1,0 +1,5 @@
+namespace Luxelot.Apps.DHT;
+
+public interface IBucketEntryValue
+{
+}