+++ apps/Luxelot.Apps.DhtApp/NodeEntry.cs
@@ -1,7 +1,8 @@
 using System.Collections.Immutable;
 using System.Net;
+using Luxelot.Apps.Common;
 
-namespace Luxelot.Apps.Common.DHT;
+namespace Luxelot.Apps.DHT;
 
 public record class NodeEntry : IBucketEntryValue
 {
@@ -9,8 +10,5 @@ public record class NodeEntry : IBucketEntryValue
 
     public required ImmutableArray<byte>? IdentityPublicKey { get; init; }
 
-    public override string ToString()
-    {
-        return $"{DisplayUtils.BytesToHex(IdentityPublicKey)[..8]}... {RemoteEndpoint}";
-    }
+    public override string ToString() => $"{DisplayUtils.BytesToHex(IdentityPublicKey)[..8]}... {RemoteEndpoint}";
 }